Se crea el dbo de employee, y conexion a la base de datos

+++ src/main/java/co/com/examen/seminario/employee/domain/model/employee/Salary.java
@@ -4,9 +4,9 @@ import lombok.Builder;
 
 @Builder
 public class Salary {
-    private Double value;
+    private Float value;
 
-    public Double getValue() {
+    public Float getValue() {
         return value;
     }
